GPS: fixed some stdint types

+++ libraries/AP_GPS/AP_GPS_NMEA.h
@@ -96,7 +96,7 @@ private:
     /// @returns		The value expressed by the string in _term,
     ///					multiplied by 100.
     ///
-    unsigned long	_parse_decimal();
+    uint32_t	_parse_decimal();
 
     /// Parses the current term as a NMEA-style degrees + minutes
     /// value with up to four decimal digits.
@@ -106,7 +106,7 @@ private:
     /// @returns		The value expressed by the string in _term,
     ///					multiplied by 10000.
     ///
-    unsigned long	_parse_degrees();
+    uint32_t	_parse_degrees();
 
     /// Processes the current term when it has been deemed to be
     /// complete.
